+++ medspacy/_extensions.py
@@ -4,7 +4,7 @@ from .common.util import span_contains
 
 # from .io.doc_consumer import ALLOWED_DATA_TYPES
 
-ALLOWED_DATA_TYPES = ("ent", "section", "context", "doc")
+ALLOWED_DATA_TYPES = ("ents", "section", "context", "doc")
 
 
 def set_extensions():